Refactor file_io
