+++ erniebot-agent/erniebot_agent/file_io/base.py
@@ -22,6 +22,12 @@ class File(metaclass=abc.ABCMeta):
         self.filename = filename
         self.created_at = created_at
 
+    def __eq__(self, other: object) -> bool:
+        if isinstance(other, File):
+            return self.id == other.id
+        else:
+            return False
+
     def __repr__(self) -> str:
         attrs_str = self._get_attrs_str()
         return f"<{self.__class__.__name__} {attrs_str}>"
